Updated the EditProfileServlet, edit.jsp, MYSQLUsersdao and Users interface to catch errors from userinput.
ISSUE: There is an issue when trying to edit the username. The error for having a duplicate username in the database occurs when swithcing the name. I will continue to troubleshoot, but I might have to eliminate the option of username editing and only allow if for the email and password.